changed camera so it doesn't have to do the opengl stuff, around 10fps now, models are better
